refactor for simplification - moved docstrings - moved method to object collections
