+++ mindsdb_sdk/query.py
@@ -1,4 +1,3 @@
-
 import pandas as pd
