+++ mindsdb_sdk/tables.py
@@ -1,13 +1,17 @@
 import copy
 from typing import Union
+from typing import List
 
 import pandas as pd
 
+from mindsdb_sql.parser.ast import DropTables
 from mindsdb_sql.parser.ast import Select, Star, Identifier, Constant, Delete, Insert, Update
 
-from mindsdb_sdk.utils import dict_to_binary_op
+from mindsdb_sdk.utils.sql import dict_to_binary_op
+from mindsdb_sdk.utils.objects_collection import CollectionBase
 
 from .query import Query
+
 
 class Table(Query):
     def __init__(self, db, name):
@@ -166,20 +170,139 @@ class Table(Query):
             raise NotImplementedError
 
 
-class View(Table):
-    # The same as table
-    pass
+class Tables(CollectionBase):
+    """
+    Wortking with tables:
+    Get table as Query object
 
-# TODO getting view sql from api not implemented yet
-# class View(Table):
-#     def __init__(self, api, data, project):
-#         super().__init__(api, data['name'], project)
-#         self.view_sql = data['sql']
-#
-#     def __repr__(self):
-#         #
-#         sql = self.view_sql.replace('\n', ' ')
-#         if len(sql) > 40:
-#             sql = sql[:37] + '...'
-#
-#         return f'{self.__class__.__name__}({self.name}{self._filters_repr()}, sql={sql})'
+    >>> table = tables.get('table1')
+
+    Filter and limit
+
+    >>> table = table.filter(a=1, b='2')
+    >>> table = table.limit(3)
+
+    Get content of table as dataframe. At that moment query will be sent on server and executed
+
+    >>> df = table.fetch()
+
+    Creating table
+
+    From query:
+
+    >>> table = tables.create('table2', query)
+
+    From other table
+
+    >>> table2 = table.create('table2', table)
+
+    Uploading file
+
+    >>> db = con.databases.files
+    >>> db.tables.create('filename', dataframe)
+
+  ` Droping table
+
+    >>> db.tables.drop('table2')
+    """
+
+    def __init__(self, database, api):
+        self.database = database
+        self.api = api
+
+    def _list_tables(self):
+        df = self.database.query('show tables').fetch()
+
+        # first column
+        return list(df[df.columns[0]])
+
+    def list(self) -> List[Table]:
+        """
+        Show list of tables in integration
+
+        :return: list of Table objects
+        """
+        return [Table(self.database, name) for name in self._list_tables()]
+
+    def get(self, name: str) -> Table:
+        """
+        Get table by name
+
+        :param name: name of table
+        :return: Table object
+        """
+
+        if name not in self._list_tables():
+            if '.' not in name:
+                # fixme: schemas not visible in 'show tables'
+                raise AttributeError("Table doesn't exist")
+        return Table(self.database, name)
+
+    def create(self, name: str, query: Union[pd.DataFrame, Query], replace: bool = False) -> Table:
+        """
+        Create new table and return it.
+
+        On mindsdb server it executes command:
+        `insert into a (select ...)`
+
+        or if replace is True
+        `create table a (select ...)`
+
+        'select ...' is extracted from input Query
+
+        :param name: name of table
+        :param query: Query object
+        :param replace: if true,
+        :return: Table object
+        """
+
+        if isinstance(query, pd.DataFrame) and self.database.name == 'files':
+            # now it is only possible for file uploading
+            self.api.upload_file(name, query)
+
+            return Table(self.database, name)
+
+        if not isinstance(query, Query):
+            raise NotImplementedError
+
+        # # query can be in different database: wrap to NativeQuery
+        # ast_query = CreateTable(
+        #     name=Identifier(name),
+        #     is_replace=is_replace,
+        #     from_select=Select(
+        #         targets=[Star()],
+        #         from_table=NativeQuery(
+        #             integration=Identifier(data.database),
+        #             query=data.sql
+        #         )
+        #     )
+        # )
+        # self.query(ast_query.to_string()).fetch()
+
+        # call in query database
+        table = Identifier(parts=[self.database.name, name])
+
+        replace_str = ''
+        if replace:
+            replace_str = ' or replace'
+
+        self.api.sql_query(
+            f'create{replace_str} table {table.to_string()} ({query.sql})',
+            database=query.database
+        )
+
+        return Table(self.database, name)
+
+    def drop(self, name: str):
+        """
+        Delete table
+
+        :param name: name of table
+        """
+        ast_query = DropTables(
+            tables=[
+                Identifier(parts=[name])
+            ]
+        )
+        self.api.sql_query(ast_query.to_string(), database=self.database.name)
+
