+++ requirements.txt
@@ -1,3 +1,3 @@
 requests
-pandas == 1.3.5
+pandas >=2.0.0
 mindsdb-sql >= 0.7.0, < 0.8.0
